refactor(codersdk): move group AI budget types into aibridge.go
